bug fix: store result in double

+++ 13_ClimbingStairs/13_ClimbingStairs.cpp
@@ -4,9 +4,9 @@
 #include <iostream>
 #include "13_ClimbingStairs.h"
 
-int Solution::fact(int n)
+double Solution::fact(int n)
 {
-	int res = 1;
+	double res = 1;
 	for (int i = 2; i <= n; i++)
 	{
 		res *= i;
@@ -22,19 +22,19 @@ int Solution::climbStairs_rec(int n)
 	return climbStairs_rec(n - 1) + climbStairs_rec(n - 2);
 }
 
-int Solution::climbStairs(int n)
+double Solution::climbStairs(int n)
 {
-	int count = 1; // with step=1
+	double count = 1; // with step=1
 	// 1 1 1 1 1 1
 	//  2  1 1 1 1 :add one 2
 	//  2   2  1 1 :add second 2
-	//  2   2   2  :ass third 3
+	//  2   2   2  :add third 2
 	// and so on...
 	for (int i = 1; i <= n / 2; i++)
 	{
 		//how many ways are there to place k twos in n places?
-		int numemator = fact(n - i);
-		int denumenator = fact((n - i) - i) * fact(i);
+		double numemator = fact(n - i);
+		double denumenator = fact((n - i) - i) * fact(i);
 		count += numemator / denumenator;
 			
 	}
